+++ platform/ixus80_sd1100/main.c
@@ -29,7 +29,7 @@ void startup()
 
 
 static const int fl_tbl[] = {6200, 7230, 8295, 9681, 11614, 14303, 18600};
-#define NUM_FL (sizeof(fl_tbl)/sizeof(fl_tbl[0]))
+#define NUM_FL (int)(sizeof(fl_tbl)/sizeof(fl_tbl[0]))
 #define CF_EFL 61291
 
 const int zoom_points = NUM_FL;
